+++ toolium/visual_test.py
@@ -16,6 +16,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
+# Python 2.7
+from __future__ import division
+
 import logging
 import os
 import re
@@ -29,12 +32,12 @@ except NameError:
     # Python 3
     xrange = range
 
-from selenium.webdriver.remote.webelement import WebElement
 from toolium import toolium_wrapper
 import itertools
 from needle.engines.perceptualdiff_engine import Engine as PerceptualEngine
 # from needle.engines.imagemagick_engine import Engine as MagickEngine
 from needle.engines.pil_engine import Engine as PilEngine
+from toolium.utils import Utils
 from PIL import Image
 
 
@@ -48,6 +51,7 @@ class VisualTest(object):
         if not self.driver_wrapper.config.getboolean_optional('VisualTests', 'enabled'):
             return
 
+        self.utils = Utils(self.driver_wrapper)
         self.logger = logging.getLogger(__name__)
         self.output_directory = toolium_wrapper.visual_output_directory
         self.baseline_directory = toolium_wrapper.visual_baseline_directory
@@ -91,8 +95,8 @@ class VisualTest(object):
             return
 
         # Search elements
-        element = self.get_element(element_or_locator)
-        exclude_elements = [self.get_element(exclude_element) for exclude_element in exclude_elements]
+        element = self.utils.get_element(element_or_locator)
+        exclude_elements = [self.utils.get_element(exclude_element) for exclude_element in exclude_elements]
 
         baseline_file = os.path.join(self.baseline_directory, '{}.png'.format(filename))
         filename_with_suffix = '{0}__{1}'.format(filename, file_suffix) if file_suffix else filename
@@ -126,25 +130,6 @@ class VisualTest(object):
             # Compare the screenshots
             self.compare_files(report_name, output_file, baseline_file, threshold)
 
-    def get_element(self, element_or_locator):
-        """Search element by its locator
-
-        :param element_or_locator: either a WebElement, a PageElement or an element locator as a tuple (locator_type,
-                                   locator_value).
-        :returns: WebElement object
-        """
-        if element_or_locator is None:
-            element = None
-        elif isinstance(element_or_locator, WebElement):
-            element = element_or_locator
-        else:
-            try:
-                # PageElement
-                element = element_or_locator.element()
-            except AttributeError:
-                element = self.driver_wrapper.driver.find_element(*element_or_locator)
-        return element
-
     def mobile_resize(self, img):
         """Resize image in iOS (native and web) and Android (web) to fit window size
 
@@ -152,26 +137,11 @@ class VisualTest(object):
         :returns: modified image object
         """
         if self.driver_wrapper.is_ios_test() or self.driver_wrapper.is_android_web_test():
-            if self.driver_wrapper.is_ios_test():
-                window_width = self.driver_wrapper.driver.get_window_size()['width']
-            else:
-                window_width = self.driver_wrapper.driver.execute_script("return window.outerWidth")
-            scale = img.size[0] / window_width
+            scale = img.size[0] / Utils(self.driver_wrapper).get_window_size()['width']
             if scale != 1:
                 new_image_size = (int(img.size[0] / scale), int(img.size[1] / scale))
                 img = img.resize(new_image_size, Image.ANTIALIAS)
         return img
-
-    def get_safari_navigation_bar_height(self):
-        """Get the height of Safari navigation bar
-
-        :returns: height of navigation bar
-        """
-        status_bar_height = 0
-        if self.driver_wrapper.is_ios_test() and self.driver_wrapper.is_web_test():
-            # ios 7.1, 8.3
-            status_bar_height = 64
-        return status_bar_height
 
     def get_element_box(self, element):
         """Get element coordinates
@@ -179,7 +149,7 @@ class VisualTest(object):
         :param element: WebElement object
         :returns: tuple with element coordinates
         """
-        offset = self.get_safari_navigation_bar_height()
+        offset = self.utils.get_safari_navigation_bar_height()
         return (int(element.location['x']), int(element.location['y'] + offset),
                 int(element.location['x'] + element.size['width']),
                 int(element.location['y'] + offset + element.size['height']))